backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Aaron <AaronBatilo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
uvicorn = "^0.13.4"
gunicorn = "^20.0.4"
uvloop = "^0.15.2"
httptools = "^0.1.1"
psycopg2-binary = "^2.8.6"
python-json-logger = "^2.0.1"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
